clp-package: Add the log-viewer and serve it using the log-viewer-webui. (#490)
