+++ Taskfile.yml
@@ -47,6 +47,7 @@ tasks:
       - "rm -rf 'components/log-viewer-webui/client/node_modules'"
       - "rm -rf 'components/log-viewer-webui/node_modules'"
       - "rm -rf 'components/log-viewer-webui/server/node_modules'"
+      - "rm -rf 'components/log-viewer-webui/yscope-log-viewer/node_modules'"
 
   clean-webui:
     cmds:
@@ -74,7 +75,7 @@ tasks:
       CHECKSUM_FILE: "{{.G_BUILD_DIR}}/{{.TASK}}.md5"
       OUTPUT_DIR: "{{.G_PACKAGE_BUILD_DIR}}"
     sources:
-      - "{{.G_BUILD_DIR}}/log-viewer-webui-client.md5"
+      - "{{.G_BUILD_DIR}}/log-viewer-webui-clients.md5"
       - "{{.G_BUILD_DIR}}/package-venv.md5"
       - "{{.G_BUILD_DIR}}/webui.md5"
       - "{{.G_BUILD_DIR}}/webui-nodejs.md5"
@@ -100,7 +101,7 @@ tasks:
       - "clp-py-utils"
       - "init"
       - "job-orchestration"
-      - "log-viewer-webui-client"
+      - "log-viewer-webui-clients"
       - "nodejs-14"
       - "package-venv"
       - task: "utils:validate-checksum"
@@ -150,6 +151,7 @@ tasks:
       - >-
         rsync -a
         "{{.G_LOG_VIEWER_WEBUI_BUILD_DIR}}/client"
+        "{{.G_LOG_VIEWER_WEBUI_BUILD_DIR}}/yscope-log-viewer"
         "{{.OUTPUT_DIR}}/var/www/log_viewer_webui/"
       - |-
         cd components/log-viewer-webui/server/
@@ -206,17 +208,23 @@ tasks:
       vars:
         COMPONENT: "{{.TASK}}"
 
-  log-viewer-webui-client:
+  log-viewer-webui-clients:
     vars:
       CHECKSUM_FILE: "{{.G_BUILD_DIR}}/{{.TASK}}.md5"
       OUTPUT_DIR: "{{.G_LOG_VIEWER_WEBUI_BUILD_DIR}}"
     sources:
-      - "{{.G_BUILD_DIR}}/log-viewer-modules.md5"
+      - "{{.G_BUILD_DIR}}/log-viewer-webui-node-modules.md5"
       - "{{.TASKFILE}}"
       - "client/package.json"
       - "client/src/**/*.css"
       - "client/src/**/*.jsx"
       - "client/src/webpack.config.js"
+      - "yscope-log-viewer/.babelrc"
+      - "yscope-log-viewer/customized-packages/**/*"
+      - "yscope-log-viewer/package.json"
+      - "yscope-log-viewer/src/**/*"
+      - "yscope-log-viewer/webpack.common.js"
+      - "yscope-log-viewer/webpack.prod.js"
     dir: "components/log-viewer-webui"
     generates: ["{{.CHECKSUM_FILE}}"]
     deps:
@@ -228,10 +236,13 @@ tasks:
           DATA_DIR: "{{.OUTPUT_DIR}}"
     cmds:
       - "rm -rf '{{.OUTPUT_DIR}}'"
-      - |-
-        cd client
-        PATH="{{.G_NODEJS_22_BIN_DIR}}":$PATH npm run build -- \
-        --output-path "{{.OUTPUT_DIR}}/client"
+      - for:
+          - "client"
+          - "yscope-log-viewer"
+        cmd: |-
+          cd "{{.ITEM}}"
+          PATH="{{.G_NODEJS_22_BIN_DIR}}":$PATH npm run build -- \
+          --output-path "{{.OUTPUT_DIR}}/{{.ITEM}}"
       - task: "utils:compute-checksum"
         vars:
           DATA_DIR: "{{.OUTPUT_DIR}}"
@@ -364,25 +375,31 @@ tasks:
           DATA_DIR: "{{.OUTPUT_DIR}}"
           OUTPUT_FILE: "{{.CHECKSUM_FILE}}"
 
-  # NOTE: The log-viewer-webui has three different node_modules directories (client, server, and the
-  # top-level one we call "package"), meaning we have to create three different checksums. To allow
-  # tasks which depend on this task to only have to check one checksum file, we concatenate the
-  # three checksum files into one.
+  # NOTE: The log-viewer-webui has four different node_modules directories:
+  # * client
+  # * server
+  # * log-viewer submodule
+  # * the top-level one we call "package"
+  # This means we have to create four different checksums. To allow tasks which depend on this task
+  # to only have to check one checksum file, we concatenate the four checksum files into one.
   log-viewer-webui-node-modules:
     internal: true
     vars:
       # Checksum files
       CHECKSUM_FILE: "{{.G_BUILD_DIR}}/{{.TASK}}.md5"
       CLIENT_CHECKSUM_FILE: "{{.G_BUILD_DIR}}/log-viewer-webui-client-node-modules.md5"
+      LOG_VIEWER_CHECKSUM_FILE: "{{.G_BUILD_DIR}}/log-viewer-webui-log-viewer-node-modules.md5"
       PACKAGE_CHECKSUM_FILE: "{{.G_BUILD_DIR}}/log-viewer-webui-package-node-modules.md5"
       SERVER_CHECKSUM_FILE: "{{.G_BUILD_DIR}}/log-viewer-webui-server-node-modules.md5"
 
       # Directories
       SRC_DIR: "{{.TASKFILE_DIR}}/components/log-viewer-webui"
       CLIENT_OUTPUT_DIR: "{{.SRC_DIR}}/client/node_modules"
+      LOG_VIEWER_OUTPUT_DIR: "{{.SRC_DIR}}/yscope-log-viewer/node_modules"
       PACKAGE_OUTPUT_DIR: "{{.SRC_DIR}}/node_modules"
       SERVER_OUTPUT_DIR: "{{.SRC_DIR}}/server/node_modules"
     sources:
+      - "{{.G_BUILD_DIR}}/log-viewer-webui-submodules.md5"
       - "{{.G_BUILD_DIR}}/nodejs-22.md5"
       - "{{.TASKFILE}}"
       - "client/package.json"
@@ -391,13 +408,17 @@ tasks:
       - "package-lock.json"
       - "server/package.json"
       - "server/package-lock.json"
+      - "yscope-log-viewer/package.json"
+      - "yscope-log-viewer/package-lock.json"
     dir: "{{.SRC_DIR}}"
     generates:
       - "{{.CHECKSUM_FILE}}"
       - "{{.CLIENT_CHECKSUM_FILE}}"
+      - "{{.LOG_VIEWER_CHECKSUM_FILE}}"
       - "{{.PACKAGE_CHECKSUM_FILE}}"
       - "{{.SERVER_CHECKSUM_FILE}}"
     deps:
+      - "log-viewer-webui-submodules"
       - "nodejs-22"
       - task: "utils:validate-checksum"
         vars:
@@ -411,15 +432,26 @@ tasks:
         vars:
           CHECKSUM_FILE: "{{.PACKAGE_CHECKSUM_FILE}}"
           DATA_DIR: "{{.PACKAGE_OUTPUT_DIR}}"
+      - task: "utils:validate-checksum"
+        vars:
+          CHECKSUM_FILE: "{{.LOG_VIEWER_CHECKSUM_FILE}}"
+          DATA_DIR: "{{.LOG_VIEWER_OUTPUT_DIR}}"
     cmds:
       - "rm -f {{.CHECKSUM_FILE}}"
       - task: "clean-log-viewer-webui"
       - "PATH='{{.G_NODEJS_22_BIN_DIR}}':$PATH npm run init"
+      - |-
+        cd yscope-log-viewer
+        PATH="{{.G_NODEJS_22_BIN_DIR}}":$PATH npm install
       # These commands must be last
       - task: "utils:compute-checksum"
         vars:
           DATA_DIR: "{{.CLIENT_OUTPUT_DIR}}"
           OUTPUT_FILE: "{{.CLIENT_CHECKSUM_FILE}}"
+      - task: "utils:compute-checksum"
+        vars:
+          DATA_DIR: "{{.LOG_VIEWER_OUTPUT_DIR}}"
+          OUTPUT_FILE: "{{.LOG_VIEWER_CHECKSUM_FILE}}"
       - task: "utils:compute-checksum"
         vars:
           DATA_DIR: "{{.PACKAGE_OUTPUT_DIR}}"
@@ -432,9 +464,34 @@ tasks:
       - >-
         cat
         "{{.CLIENT_CHECKSUM_FILE}}"
+        "{{.LOG_VIEWER_CHECKSUM_FILE}}"
         "{{.PACKAGE_CHECKSUM_FILE}}"
         "{{.SERVER_CHECKSUM_FILE}}"
         > "{{.CHECKSUM_FILE}}"
+
+  log-viewer-webui-submodules:
+    internal: true
+    vars:
+      CHECKSUM_FILE: "{{.G_BUILD_DIR}}/{{.TASK}}.md5"
+      OUTPUT_DIR: "yscope-log-viewer"
+    sources:
+      - "{{.TASKFILE}}"
+      - ".gitmodules"
+    dir: "components/log-viewer-webui"
+    generates: ["{{.CHECKSUM_FILE}}"]
+    deps:
+      - "init"
+      - task: "utils:validate-checksum"
+        vars:
+          CHECKSUM_FILE: "{{.CHECKSUM_FILE}}"
+          DATA_DIR: "{{.OUTPUT_DIR}}"
+    cmds:
+      - "git submodule update --init --recursive yscope-log-viewer"
+      # This command must be last
+      - task: "utils:compute-checksum"
+        vars:
+          DATA_DIR: "{{.OUTPUT_DIR}}"
+          OUTPUT_FILE: "{{.CHECKSUM_FILE}}"
 
   meteor:
     vars:
