+++ components/clp-package-utils/clp_package_utils/scripts/start_clp.py
@@ -864,6 +864,7 @@ def start_log_viewer_webui(
         "MongoDbIrFilesCollectionName": clp_config.results_cache.ir_collection_name,
         "ClientDir": str(container_log_viewer_webui_dir / "client"),
         "IrFilesDir": str(container_clp_config.ir_output.directory),
+        "LogViewerDir": str(container_log_viewer_webui_dir / "yscope-log-viewer"),
     }
     settings_json = read_and_update_settings_json(settings_json_path, settings_json_updates)
     with open(settings_json_path, "w") as settings_json_file:
