+++ components/log-viewer-webui/server/settings.json
@@ -9,5 +9,6 @@
     "MongoDbIrFilesCollectionName": "ir-files",
 
     "ClientDir": "../client/dist",
-    "IrFilesDir": "../../../build/clp-package/var/data/ir"
+    "IrFilesDir": "../../../build/clp-package/var/data/ir",
+    "LogViewerDir": "../yscope-log-viewer/dist"
 }
